fully working tests

+++ apps/Interactions/apis.py
@@ -3,6 +3,7 @@ from fastapi.responses import JSONResponse
 from sqlmodel import select, func
 from uuid import UUID, uuid4
 
+from apps.Clients.apis import UserRole
 from apps.Interactions.models import (
     Interaction,
     InteractionCreate,
@@ -14,16 +15,17 @@ from config.settings import settings
 from apps.Users.models import User
 from apps.deps import SessionDep, get_current_user
 
-#TODO вынести зависимость гет каррент юзер в deps
 router = APIRouter(prefix="/interactions", tags=["interaction"])
 
 @router.get("/", response_model=InteractionList)
 def get_interactions(
     session: SessionDep,
-    # user: User = Depends(get_current_user),
+    current_user: User = Depends(get_current_user),
     skip: int = 0,
     limit: int = 100
 ) -> InteractionPublic:
+    if current_user.role != UserRole.admin:
+        raise HTTPException(detail="Not enough permission", status_code=status.HTTP_400_BAD_REQUEST)
     limit = min(limit, settings.MAX_LIMIT)
     interactions = session.exec(select(Interaction).order_by(Interaction.created_at).offset(skip).limit(limit))
     count = session.exec(select(func.count()).select_from((Interaction))).one()
@@ -37,13 +39,15 @@ def get_interactions(
 def get_interaction(
     session: SessionDep,
     id: UUID,
-    # user: User = Depends(get_current_user)
+    current_user: User = Depends(get_current_user)
 ) -> InteractionPublic:
     interaction = session.get(Interaction, id)
     if not interaction:
         raise HTTPException(detail="Interaction not found", status_code=status.HTTP_404_NOT_FOUND)
     if not interaction.is_active:
         raise HTTPException(detail="Interaction is inactive", status_code=status.HTTP_409_CONFLICT)
+    if current_user.id != interaction.user_id and current_user.role != UserRole.admin:
+        raise HTTPException(detail="Not enough permission", status_code=status.HTTP_400_BAD_REQUEST)
     return interaction
 
 
@@ -51,7 +55,7 @@ def get_interaction(
 def create_interaction(
     session: SessionDep,
     interaction_in: InteractionCreate,
-    # user: User = Depends(get_current_user),
+    current_user: User = Depends(get_current_user),
 ) -> InteractionPublic:
     data = interaction_in.model_dump()
     interaction = Interaction(
@@ -69,13 +73,13 @@ def update_interaction(
     session: SessionDep,
     interaction_in: InteractionUpdate,
     id: UUID,
-    # user: User = Depends(get_current_user)
+    current_user: User = Depends(get_current_user)
 ) -> InteractionPublic:
     interaction = session.get(Interaction, id)
     if not interaction:
         raise HTTPException(detail="Interaction not found", status_code=status.HTTP_404_NOT_FOUND)
-    if not interaction.is_active:
-        raise HTTPException(detail="Interaction is inactive", status_code=status.HTTP_409_CONFLICT)
+    if current_user.id != interaction.user_id and current_user.role != UserRole.admin:
+        raise HTTPException(detail="Not enough permission", status_code=status.HTTP_400_BAD_REQUEST)
     data = interaction_in.model_dump(exclude_unset=True)
     interaction.sqlmodel_update(data)
     session.add(interaction)
@@ -88,11 +92,13 @@ def update_interaction(
 def delete_interaction(
     session: SessionDep,
     id: UUID,
-    # user: User = Depends(get_current_user)
+    current_user: User = Depends(get_current_user)
 ) -> JSONResponse:
     interaction = session.get(Interaction, id)
     if not interaction:
         raise HTTPException(detail="Interaction not found", status_code=status.HTTP_404_NOT_FOUND)
+    if current_user.id != interaction.user_id and current_user.role != UserRole.admin:
+        raise HTTPException(detail="Not enough permission", status_code=status.HTTP_400_BAD_REQUEST)
     session.delete(interaction)
     session.commit()
     return JSONResponse(content={"detail": "Interaction was deleted successfully"}, status_code=status.HTTP_200_OK)
